update action to only support one history at a time
