--- extensions/elation/actions/addHistories/index.ts
@@ -1,1 +1,0 @@
-export { addHistories } from './addHistories'
